Refactor ContentView into modular view files
Split ContentView.swift (732 lines) into focused, maintainable files:

New structure:
- ContentView.swift (46 lines) - Main container with state routing
- SessionIntroView.swift - IntroView
- PreWorkoutFlowViews.swift - PreWorkoutView, StretchingView
- WarmupView.swift - WarmupView
- WorkingSetViews.swift - LoadView, WorkingSetView, RestView
- SessionReviewViews.swift - ReviewView, DoneView
- Components/ directory:
  - ExerciseHeader.swift
  - ActionBar.swift
  - DecisionBadge.swift

Benefits:
- Reduced ContentView from 732 to 46 lines (93% reduction)
- Better code organization by workout phase
- Improved navigation and maintainability
- Faster compile times with smaller compilation units
- Clear separation of concerns

All views maintain identical functionality. Build verified successful.

Fixes #12
